Iterators: get_center, get_com
